+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/statistics/UsageStatisticsReportsSettingsDeterminer.scala
@@ -1,18 +1,23 @@
 package pl.touk.nussknacker.ui.statistics
 
 import io.circe.generic.JsonCodec
+import org.apache.commons.io.FileUtils
+import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
 import pl.touk.nussknacker.engine.version.BuildInfo
 import pl.touk.nussknacker.restmodel.process.ProcessingType
 import pl.touk.nussknacker.ui.config.UsageStatisticsReportsConfig
 import pl.touk.nussknacker.ui.process.processingtypedata.ProcessingTypeDataProvider
+import pl.touk.nussknacker.ui.statistics.UsageStatisticsReportsSettingsDeterminer._
 
+import java.io.File
 import java.net.URLEncoder
 import java.nio.charset.StandardCharsets
 import scala.collection.immutable.ListMap
-import scala.util.Random
-import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
+import scala.util.{Random, Try}
 
-object UsageStatisticsReportsSettings {
+object UsageStatisticsReportsSettingsDeterminer {
+
+  private val nuFingerprintFileName = "nussknacker.fingerprint"
 
   private val knownDeploymentManagerTypes = Set("flinkStreaming", "lite-k8s", "lite-embedded")
 
@@ -23,21 +28,43 @@ object UsageStatisticsReportsSettings {
   // We aggregate custom deployment managers and processing modes as a "custom" to avoid leaking of internal, confidential data
   private val aggregateForCustomValues = "custom"
 
-  def prepare(config: UsageStatisticsReportsConfig, processingTypeStatistics: ProcessingTypeDataProvider[ProcessingTypeUsageStatistics, _]): UsageStatisticsReportsSettings = {
-    val queryParams = prepareQueryParams(config, processingTypeStatistics.all)
+  def apply(config: UsageStatisticsReportsConfig,
+            processingTypeStatistics: ProcessingTypeDataProvider[ProcessingTypeUsageStatistics, _]): UsageStatisticsReportsSettingsDeterminer = {
+    UsageStatisticsReportsSettingsDeterminer(config, processingTypeStatistics.all)
+  }
+
+  def apply(config: UsageStatisticsReportsConfig,
+            processingTypeStatistics: => Map[ProcessingType, ProcessingTypeUsageStatistics]): UsageStatisticsReportsSettingsDeterminer = {
+    val fingerprintFile = new File(Try(Option(System.getProperty("java.io.tmpdir"))).toOption.flatten.getOrElse("/tmp"), nuFingerprintFileName)
+    new UsageStatisticsReportsSettingsDeterminer(config, processingTypeStatistics, fingerprintFile)
+  }
+
+  private[statistics] def prepareUrl(queryParams: ListMap[String, String]) = {
+    queryParams.toList.map {
+      case (k, v) => s"${URLEncoder.encode(k, StandardCharsets.UTF_8)}=${URLEncoder.encode(v, StandardCharsets.UTF_8)}"
+    }.mkString("https://stats.nussknacker.io/?", "&", "")
+  }
+
+}
+
+class UsageStatisticsReportsSettingsDeterminer(config: UsageStatisticsReportsConfig,
+                                               processingTypeStatistics: => Map[ProcessingType, ProcessingTypeUsageStatistics],
+                                               fingerprintFile: File) {
+
+  def determineSettings(): UsageStatisticsReportsSettings = {
+    val queryParams = determineQueryParams()
     val url = prepareUrl(queryParams)
     UsageStatisticsReportsSettings(config.enabled, url)
   }
 
-  private[statistics] def prepareQueryParams(config: UsageStatisticsReportsConfig,
-                                             processingTypeStatisticsMap: Map[ProcessingType, ProcessingTypeUsageStatistics]): ListMap[String, String] = {
-    val deploymentManagerTypes = processingTypeStatisticsMap.values.map(_.deploymentManagerType).map {
+  private[statistics] def determineQueryParams(): ListMap[String, String] = {
+    val deploymentManagerTypes = processingTypeStatistics.values.map(_.deploymentManagerType).map {
       case Some(dm) if knownDeploymentManagerTypes.contains(dm) => dm
       case _ => aggregateForCustomValues
     }
     val dmParams = prepareValuesParams(deploymentManagerTypes, "dm")
 
-    val processingModes = processingTypeStatisticsMap.values.map {
+    val processingModes = processingTypeStatistics.values.map {
       case ProcessingTypeUsageStatistics(_, Some(mode)) if knownProcessingModes.contains(mode) => mode
       case ProcessingTypeUsageStatistics(Some(deploymentManagerType), None) if deploymentManagerType.toLowerCase.contains(streamingProcessingMode) => streamingProcessingMode
       case _ => aggregateForCustomValues
@@ -47,7 +74,7 @@ object UsageStatisticsReportsSettings {
     ListMap(
       // We filter out blank fingerprint and source because when smb uses docker-compose, and forwards env variables eg. USAGE_REPORTS_FINGERPRINT
       // from system and the variable doesn't exist, there is no way to skip variable - it can be only set to empty
-      "fingerprint" -> config.fingerprint.filterNot(_.isBlank).getOrElse(randomFingerprint),
+      "fingerprint" -> config.fingerprint.filterNot(_.isBlank).getOrElse(fingerprint),
       // If it is not set, we assume that it is some custom build from source code
       "source" -> config.source.filterNot(_.isBlank).getOrElse("sources"),
       "version" -> BuildInfo.version
@@ -67,13 +94,19 @@ object UsageStatisticsReportsSettings {
     ListMap(countsParams: _*) + (s"single_$metricCategoryKeyPart" -> singleParamValue)
   }
 
-  private[statistics] def prepareUrl(queryParams: ListMap[String, String]) = {
-    queryParams.toList.map {
-      case (k, v) => s"${URLEncoder.encode(k, StandardCharsets.UTF_8)}=${URLEncoder.encode(v, StandardCharsets.UTF_8)}"
-    }.mkString("https://stats.nussknacker.io/?", "&", "")
+  private lazy val fingerprint: String = persistedFingerprint {
+    randomFingerprint
   }
 
-  private lazy val randomFingerprint = s"gen-${Random.alphanumeric.take(10).mkString}"
+  private def persistedFingerprint(compute: => String) = {
+    Try(FileUtils.readFileToString(fingerprintFile, StandardCharsets.UTF_8)).getOrElse {
+      val f = compute
+      Try(FileUtils.writeStringToFile(fingerprintFile, f, StandardCharsets.UTF_8))
+      f
+    }
+  }
+
+  private def randomFingerprint = s"gen-${Random.alphanumeric.take(10).mkString}"
 
 }